+++ src/main/java/com/example/Proveedores_Empresariales/Company/CompanyService.java
@@ -3,6 +3,7 @@ package com.example.Proveedores_Empresariales.Company;
 import com.example.Proveedores_Empresariales.City.City;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -12,9 +13,9 @@ public interface CompanyService {
 
     public List<Company> getAll();
 
-    public Company getById(int id);
+    public Company getById(BigInteger id);
 
-    public Company update(int nic,Company company);
+    public Company update(BigInteger nic,Company company);
 
     public void delete (Company company);
 }
