+++ src/main/java/com/example/Proveedores_Empresariales/Company/RepositoryCompany.java
@@ -3,5 +3,7 @@ package com.example.Proveedores_Empresariales.Company;
 import com.example.Proveedores_Empresariales.City.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RepositoryCompany extends JpaRepository<Company,Integer> {
+import java.math.BigInteger;
+
+public interface RepositoryCompany extends JpaRepository<Company, BigInteger> {
 }
